Refactor Docker deployment
This change does a whole bunch of refactoring for running Nadekobot
in Docker.

- Database and data are now two separate volumes. That way, the data
volume can be deleted whenever there are data updates without affecting
the db.
- The Dockerfile facilities caching of the project dependencies.
However, changes to data still trigger a rebuild of the code because of
how the project files are set up.
- The project file was updated to use .NET Core 3.1, which was already
required by the install scripts for Linux. This fixes some dependency
issues with dotnet restore in the Docker build.