Fix version mismatch for transitive dependency
The dependency chain for this is
1. linq2db.EntityFrameworkCore
2. Microsoft.EntityFrameworkCore.Relational
3. Microsoft.EntityFrameworkCore
4. Microsoft.Bcl.AsyncInterfaces

However, we already have an explicit dependency on
Microsoft.EntityFrameworkCore. To prevent a version mismatch between
these two, we specify a lower version of linq2db.EntityFrameworkCore so
that it doesn't pull in a newer version of
Microsoft.Bcl.AsyncInterfaces that is incompatible with the one we have
already.