BLACKJACK CHANGE TO CLASSICAL BLACKJACK
Description
Blackjack now pushes when dealer.value = player.value, and on Blackjack (playervalue = 21) pays out 5:2 instead of 2:1

Changes Proposed

Adds "Tie" UserState in /blackjack/player
Changed conditional operators to if statements for clarity in Blackjack.cs
Added blackjack "blackjack" win and blackjack push in addition to normal blackjack win as a response.

Details
Clarity for the ifstatements, and new responses for new actions. This is closer to classical blackjack.