Add hacky fix for data updates
Data updates work by baking the "default" files into the container and
copying them over with rsync in the startup script into the real volume.
This commit adds a hack to make rsync bring in updated data files.
Unfortunately, if the default for a config file is updated, this will
overwrite existing configs.